Split up Visitor.cc and Visitor.h
Use one file per class. Also, use boost::core::demangle instead of
checking for abi::__cxa_demangle by ourselves.
