+++ example/demoApp.cc
@@ -100,6 +100,8 @@ void ExampleApp::defineConnections() {
     dev("probeSignal", typeid(int), tableLength) [ macropulseNr ] >> cs("probeSignal");
     
     dumpConnections();
-
+    dumpConnectionGraph();
+    dumpGraph();
+    dumpModuleGraph("module-graph.dot");
 }
