+++ include/Module.h
@@ -111,6 +111,7 @@ namespace ChimeraTK {
 
       /**
        * Explcitly add accept() method so that we can distinguish between a Module and an EntityOwner in the Visitor.
+
        */
       void accept(Visitor<Module>& visitor) const { visitor.dispatch(*this); }
