+++ include/VisitorHelper.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <string>
+
+namespace ChimeraTK {
+class VariableNetworkNode;
+
+namespace detail {
+
+std::string encodeDotNodeName(std::string name);
+std::string nodeName(const VariableNetworkNode& node);
+
+} // namespace detail)
+}// namespace ChimeraTK
